Revert "Trying to have one form for add and edit. NOT WORKING"
This reverts commit f596c2d725477fce1cbdd6b195f5a2d1841f4b15.
