+++ src/app/features/plants/edit-plant.tsx
@@ -3,10 +3,7 @@ import { ProgressBar } from "@/app/features/progress-bar";
 import { PlantForm } from "@/app/features/plants/plant-form";
 import useSWR from "swr";
 import { useParams } from "next/navigation";
-import { formSchemaSubmit, PlantSchemaGetSingle } from "@/app/lib/schema";
-import { useForm } from "react-hook-form";
-import * as z from "zod";
-import { zodResolver } from "@hookform/resolvers/zod";
+import { PlantSchemaGetSingle } from "@/app/lib/schema";
 
 interface PlantProps {
   plant: PlantSchemaGetSingle;
@@ -17,23 +14,6 @@ export function EditPlant() {
 
   const { id } = useParams<{ id: string }>();
   const { data, error, isLoading } = useSWR(`/api/plants/${id}`, fetcher);
-  let plant = data.data;
-
-  const form = useForm<z.infer<typeof formSchemaSubmit>>({
-    resolver: zodResolver(formSchemaSubmit),
-    defaultValues: {
-      // to populate form, add data in default values
-      name: plant.name ? plant.name : "",
-      location: plant.location ? plant.location : "Select plant location",
-      frequency: plant.frequency ? plant.frequency : "0",
-      status: plant.status ? plant.status : "needs watering",
-      start: plant.start ? plant.start : null,
-      end: plant.end ? plant.end : null,
-      volume: plant.volume ? plant.volume : "Select volume of water",
-      instructions: plant.instructions ? plant.instructions : "",
-      photo: plant.photo ? plant.photo : "insert photo of plant here",
-    },
-  });
 
   if (isLoading) {
     return (
@@ -46,7 +26,7 @@ export function EditPlant() {
   } else {
     return (
       <div>
-        <PlantForm plant={plant} form={form} />
+        <PlantForm plant={data.data} />
       </div>
     );
   }
